Various changes preparing for 0.2.0
introduce env at jsExports
handle growable memory
introduce is32Bit to constrain long to int
introduce toOpt to safely transform a T to Optional!(T*)
have bindings accept both Optional!(T) and Optional!(T*)
change P().as!(T) to return T* to &P
move to Typed allocator pools
handle growing memory not being continuous
call destructors when GC is freeing garbage
implement _d_allocmemory to support closures
introduce static create function on TypedArray's (like Float32Array)

+++ source/spasm/bindings/common.d
@@ -48,7 +48,7 @@ struct DOMException {
   enum ushort INVALID_NODE_TYPE_ERR = 24;
   enum ushort DATA_CLONE_ERR = 25;
 }
-alias DOMTimeStamp = ulong;
+alias DOMTimeStamp = uint;
 alias Function = Any delegate(Any);
 alias VoidFunction = void delegate();
